[Tests][IR] Move tailrec diagnostics testdata to testsWithAnyBackend
^KT-87442 Fixed
