+++ compiler/testData/diagnostics/tests/overload/noConflictingOverloadsWithDeprecatedHidden/FunctionsAndConstructorsInPackage.kt
@@ -1,5 +1,5 @@
 // RUN_PIPELINE_TILL: BACKEND
-// DIAGNOSTICS: -NOTHING_TO_INLINE, -NO_TAIL_CALLS_FOUND, -MISPLACED_TYPE_PARAMETER_CONSTRAINTS
+// DIAGNOSTICS: -NOTHING_TO_INLINE, -MISPLACED_TYPE_PARAMETER_CONSTRAINTS
 
 package pkg
 
@@ -66,17 +66,6 @@ class TestInlineFunctionVsConstructorReverse {
     constructor()
 }
 @Deprecated(message = "", level = DeprecationLevel.HIDDEN) inline fun TestInlineFunctionVsConstructorReverse() {}
-
-
-class TestTailrecFunctionVsConstructor {
-    @Deprecated(message = "", level = DeprecationLevel.HIDDEN) constructor()
-}
-tailrec fun TestTailrecFunctionVsConstructor() {}
-
-class TestTailrecFunctionVsConstructorReverse {
-    constructor()
-}
-@Deprecated(message = "", level = DeprecationLevel.HIDDEN) tailrec fun TestTailrecFunctionVsConstructorReverse() {}
 
 
 class TestFunctionVsPrimaryConstructor @Deprecated(message = "", level = DeprecationLevel.HIDDEN) constructor()
@@ -433,5 +422,4 @@ interface UserInterfaceA
 interface UserInterfaceB
 
 /* GENERATED_FIR_TAGS: classDeclaration, functionDeclaration, inline, interfaceDeclaration, nullableType,
-primaryConstructor, reified, secondaryConstructor, stringLiteral, tailrec, typeAliasDeclaration, typeConstraint,
-typeParameter */
+primaryConstructor, reified, secondaryConstructor, stringLiteral, typeAliasDeclaration, typeConstraint, typeParameter */
