+++ compiler/testData/diagnostics/testsWithAnyBackend/tailrec/tailRecBasic.kt
