+++ compiler/testData/diagnostics/testsWithAnyBackend/tailrec/tailRecInNestedScopes.kt
