+++ compiler/testData/diagnostics/testsWithAnyBackend/tailrec/tailRecInTry.kt
