+++ compiler/testData/diagnostics/testsWithAnyBackend/tailrec/tailRecOnVirtualMemberError.kt
