+++ compiler/testData/diagnostics/testsWithAnyBackend/tailrec/tailRecShortCircuit.kt
