+++ compiler/testData/diagnostics/testsWithAnyBackend/tailrec/tailRecSingleton.kt
