+++ compiler/testData/diagnostics/testsWithAnyBackend/tailrec/tailRecWithDispatchReceiver.kt
