+++ compiler/testData/diagnostics/testsWithAnyBackend/tailrec/tailRecWithExtensionReceiver.kt
