+++ compiler/testData/diagnostics/testsWithAnyBackend/tailrec/tailrec.kt
